R computation specific classes moved to the proper package uk.ac.ebi.gxa.R.compute
