+++ atlas-analytics/src/main/java/uk/ac/ebi/gxa/R/compute/ComputeTask.java
@@ -20,7 +20,7 @@
  * http://gxa.github.com/gxa
  */
 
-package uk.ac.ebi.gxa.analytics.compute;
+package uk.ac.ebi.gxa.R.compute;
 
 import uk.ac.ebi.rcloud.server.RServices;
