+++ atlas-analytics/src/main/java/uk/ac/ebi/gxa/R/compute/rcommand/RCommandResult.java
@@ -1,4 +1,4 @@
-package ae3.service.experiment.rcommand;
+package uk.ac.ebi.gxa.R.compute.rcommand;
 
 import uk.ac.ebi.rcloud.server.RType.*;
