+++ atlas-analytics/src/main/java/uk/ac/ebi/gxa/R/compute/rcommand/RCommandStatement.java
@@ -1,4 +1,4 @@
-package ae3.service.experiment.rcommand;
+package uk.ac.ebi.gxa.R.compute.rcommand;
 
 import com.google.common.base.Function;
