+++ atlas-analytics/src/test/java/uk/ac/ebi/gxa/R/compute/RFactoryTest.java
@@ -20,7 +20,7 @@
  * http://gxa.github.com/gxa
  */
 
-package uk.ac.ebi.gxa.analytics.compute;
+package uk.ac.ebi.gxa.R.compute;
 
 import org.junit.*;
 import org.junit.runner.RunWith;
