+++ atlas-analytics/src/test/java/uk/ac/ebi/gxa/R/compute/rcommand/RCommandStatementTest.java
@@ -1,4 +1,4 @@
-package ae3.service.experiment.rcommand;
+package uk.ac.ebi.gxa.R.compute.rcommand;
 
 import org.junit.Test;
