+++ atlas-analytics/src/test/resources/biocep.properties
