+++ atlas-analytics/src/test/resources/uk/ac/ebi/gxa/R/compute/rcommand/RCommandResultTest.R
