+++ atlas-loader/src/main/java/uk/ac/ebi/gxa/loader/steps/HTSArrayDataStep.java
@@ -10,9 +10,9 @@ import uk.ac.ebi.arrayexpress2.magetab.datamodel.sdrf.node.ScanNode;
 import uk.ac.ebi.arrayexpress2.magetab.datamodel.sdrf.node.SourceNode;
 import uk.ac.ebi.arrayexpress2.magetab.datamodel.sdrf.node.attribute.CharacteristicsAttribute;
 import uk.ac.ebi.arrayexpress2.magetab.utils.MAGETABUtils;
-import uk.ac.ebi.gxa.analytics.compute.AtlasComputeService;
-import uk.ac.ebi.gxa.analytics.compute.ComputeTask;
-import uk.ac.ebi.gxa.analytics.compute.RUtil;
+import uk.ac.ebi.gxa.R.compute.AtlasComputeService;
+import uk.ac.ebi.gxa.R.compute.ComputeTask;
+import uk.ac.ebi.gxa.R.compute.RUtil;
 import uk.ac.ebi.gxa.loader.AtlasLoaderException;
 import uk.ac.ebi.gxa.loader.cache.AtlasLoadCache;
 import uk.ac.ebi.gxa.loader.dao.LoaderDAO;
